Move pager to ui folder
