+++ js/renovation/ui/pager/__tests__/info.test.tsx
