+++ js/renovation/ui/pager/common/consts.ts
