+++ js/renovation/ui/pager/common/light_button.tsx
@@ -1,10 +1,10 @@
 import {
   Component, ComponentBindings, JSXComponent, OneWay, Slot, Event, Ref, Effect,
 } from 'devextreme-generator/component_declaration/common';
-import { registerKeyboardAction } from '../../../ui/shared/accessibility';
+import { registerKeyboardAction } from '../../../../ui/shared/accessibility';
 import { PAGER_CLASS } from './consts';
 import { closestClass } from '../utils/closest_class';
-import { subscribeToClickEvent } from '../../utils/subscribe_to_event';
+import { subscribeToClickEvent } from '../../../utils/subscribe_to_event';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const viewFunction = ({
