+++ js/renovation/ui/pager/content.tsx
@@ -9,7 +9,7 @@ import { PageIndexSelector } from './pages/page_index_selector';
 import { PageSizeSelector } from './page_size/selector';
 import { PAGER_PAGES_CLASS, LIGHT_MODE_CLASS, PAGER_CLASS } from './common/consts';
 import PagerProps from './common/pager_props';
-import { combineClasses } from '../utils/combine_classes';
+import { combineClasses } from '../../utils/combine_classes';
 
 const STATE_INVISIBLE_CLASS = 'dx-state-invisible';
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
