+++ js/renovation/ui/pager/pages/__tests__/page_index_selector.test.tsx
