+++ js/renovation/ui/pager/pages/small.tsx
@@ -11,8 +11,8 @@ import {
 
 import { Page } from './page';
 import { PAGER_INFO_CLASS } from '../info';
-import { NumberBox } from '../../ui/number_box';
-import messageLocalization from '../../../localization/message';
+import { NumberBox } from '../../number_box';
+import messageLocalization from '../../../../localization/message';
 import { calculateValuesFittedWidth } from '../utils/calculate_values_fitted_width';
 import { getElementMinWidth } from '../utils/get_element_width';
