+++ js/renovation/ui/pager/utils/calculate_values_fitted_width.ts
