+++ js/renovation/ui/pager/utils/get_computed_style.ts
