+++ js/renovation/ui/pager/utils/get_element_width.ts
